fix issue with double quotes in encrypted strings
If one of the value strings in the input object contains the character ", it raised an "Unexpected identifier" error. It was due to the function eval which evaluated expressions such as : foo = ""bar""
The expression made with concatenations is now constructed with a JSON.stringify, which escapes the double quotes.